feat: warn if binaries or configuration files are missing
Until now pipedput was rather quiet if binaries or configuration files
that are required to execute hooks were missing. This made it hard to
spot configuration errors at configuration time.

From now on pipedput will warn if the bins/files are missing and will
report a more helpful error message in report mails, if a deployment
failed due to the aforementioned circumstances.