feat: ease publication to internal GitLab PyPI project repositories
GitLab comes with its own PyPI-style project repositories for some time
now. It was already possible to publish to them through the use of a
per-project pypirc file. Unfortunately pipedput required a separate
configuration file for each project, because the repository URL is
unique per project, even though the credentials for accessing the
repository might be the same.

The PublishToPythonRepository hook can now be called with the
`publish_to_gitlab` keyword argument. This injects the correct
per-project repository URL into the call to twine and allows users to
specify a single .pypirc with only credentials.